Factorize mapper and entities
